Updated test structure and created a more clear file structure for the errors
