+++ test/module/core/errors/type-errors/readonly-write.ts
@@ -1,0 +1,30 @@
+import { KipperCompiler, KipperCompileResult, KipperError } from "@kipper/core";
+import { defaultConfig, ensureErrorWasReported, ensureTracebackDataExists } from "../index";
+import { assert } from "chai";
+
+describe("ReadOnlyWriteTypeError", () => {
+	it("Error", async () => {
+		let result: KipperCompileResult | undefined = undefined;
+		try {
+			result = await new KipperCompiler().compile(`const invalid: str = "3"; invalid = "5";`, defaultConfig);
+		} catch (e) {
+			assert.equal((<KipperError>e).constructor.name, "ReadOnlyWriteTypeError", "Expected proper error");
+			assert((<KipperError>e).name === "TypeError", "Expected proper error");
+			ensureErrorWasReported(typeof result === "object" ? result?.programCtx : undefined);
+			ensureTracebackDataExists(<KipperError>e);
+			return;
+		}
+		assert.fail("Expected 'ReadOnlyWriteTypeError'");
+	});
+
+	it("NoError", async () => {
+		let result: KipperCompileResult | undefined = undefined;
+		try {
+			result = await new KipperCompiler().compile('var valid: str = "3"; valid = "5";', defaultConfig);
+		} catch (e) {
+			assert.fail(`Expected no '${(<KipperError>e).name}'`);
+		}
+		assert(result !== undefined, "Expected result");
+		assert.isFalse(result?.programCtx.hasFailed ?? false, "Expected no errors");
+	});
+});
