+++ test/module/core/errors/type-errors/unknown-type.ts
@@ -1,0 +1,20 @@
+import { KipperCompiler, KipperCompileResult, KipperError } from "@kipper/core";
+import { defaultConfig, ensureErrorWasReported, ensureTracebackDataExists } from "../index";
+import { assert } from "chai";
+
+describe("UnknownTypeError", () => {
+	["x", "number", "UNKNOWN"].forEach((typeName) => {
+		it(typeName, async () => {
+			let result: KipperCompileResult | undefined = undefined;
+			try {
+				result = await new KipperCompiler().compile(`var invalid: ${typeName} = 4;`, defaultConfig);
+			} catch (e) {
+				assert.equal((<KipperError>e).constructor.name, "UnknownTypeError", "Expected proper error");
+				ensureErrorWasReported(typeof result === "object" ? result?.programCtx : undefined);
+				ensureTracebackDataExists(<KipperError>e);
+				return;
+			}
+			assert.fail("Expected 'UnknownTypeError'");
+		});
+	});
+});
